pyproject.toml
[tool.poetry]
name = "mykaggle"
version = "0.1.0"
description = ""
authors = ["cfiken <cfiken@gmail.com>"]

[tool.poetry.dependencies]
python = "=3.10.4"
jupyter = "^1.0.0"
jupyterlab = "^4.0.2"
numpy = "^1.24.3"
matplotlib = "3.7.1"
pillow = "^9.5.0"
pandas = "^2.0.2"
seaborn = "^0.12.2"
lightgbm = "^3.3.5"
xgboost = "^1.7.5"
catboost = "^1.2"
fastprogress = "^1.0.3"
kaggle = "^1.5.13"
pyyaml = "^6.0"
mlflow = "^2.4.1"
python-dotenv = "^0.15.0"
boto3 = "1.26.152"
types-PyYAML = "^5.4.6"
types-requests = "^2.25.6"
category-encoders = "^2.6.1"
protobuf = "^3.20.0"
opencv-python = "^4.7.0.72"
pyarrow = "^8.0.0"
datasets = "^2.3.2"
fsspec = "^2022.5.0"
torch = "^2.0.0"
albumentations = "^1.3.1"
timm = "^0.9.2"
transformers = "^4.30.1"
sentencepiece = "^0.1.96"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
mypy = "*"
autopep8 = "*"

[tool.mypy]
ignore_missing_imports = true
exclude = '^jupyter_notebook_config\.py$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
